+++ RTSD_Taecho/Assets/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4beded0afc4d8494c935e014a74c545f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
