+++ RTSD_Taecho/Assets/Fonts/Galmuri11 SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fb4c16463bd074449a162f3956c048e9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
