+++ RTSD_Taecho/Assets/Prefabs/BuildableTile.prefab
@@ -11,9 +11,10 @@ GameObject:
   - component: {fileID: 6517213697133885730}
   - component: {fileID: 2406365455620228725}
   - component: {fileID: 5218098031467866239}
+  - component: {fileID: 3127544781342708673}
   m_Layer: 0
   m_Name: BuildableTile
-  m_TagString: Untagged
+  m_TagString: Tile
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -106,3 +107,15 @@ SpriteRenderer:
   m_WasSpriteAssigned: 0
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
+--- !u!114 &3127544781342708673
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3849776298096249874}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7d17f9b7426e0ee469854b310db82270, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
