+++ RTSD_Taecho/Assets/Resource/Sound/Undead Survivor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5cb92746c6baa4742aabad8ef4fbd3ec
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
