+++ RTSD_Taecho/Assets/Resource/Sound/Undead Survivor/Audio.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1a7c796915fc4a43886391bbf749b15
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
