+++ RTSD_Taecho/Assets/Scripts/Tower/CheckDuplication.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d17f9b7426e0ee469854b310db82270
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
