+++ RTSD_Taecho/Assets/Scripts/Tower/ObjectDetector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c83828bc6c0a4344baebcd50aecfe428
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
