+++ RTSD_Taecho/ProjectSettings/TagManager.asset
@@ -10,6 +10,7 @@ TagManager:
   - T_L_3
   - T_L_4
   - T_L_5
+  - Tile
   layers:
   - Default
   - TransparentFX
